+++ src/__tests__/EditableText.test.js
@@ -1,9 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { shallow, mount } from 'enzyme';
+import keycode from 'keycode';
 
 import EditableText, { PureEditableText } from '../EditableText';
 import { PureText } from '../Text';
+
+function delayForMilliseconds(millisecond) {
+    return new Promise(resolve => setTimeout(resolve, millisecond));
+}
 
 describe('withStatus(EditableText)', () => {
     it('renders without crashing', () => {
@@ -89,6 +94,16 @@ describe('pure <PureEditableText>', () => {
         expect(input.prop('value')).toBe('Bar');
     });
 
+    it('does not clear cached input value when goes from controlled input to uncontrolled', () => {
+        const wrapper = mount(<PureEditableText value="Foo" />);
+        const input = wrapper.find('input');
+
+        expect(input.prop('value')).toBe('Foo');
+
+        wrapper.setProps({ value: null });
+        expect(input.prop('value')).toBe('Foo');
+    });
+
     it('can have a default value', () => {
         const wrapper = mount(<PureEditableText defaultValue="Foo" />);
         const input = wrapper.find('input');
@@ -99,5 +114,78 @@ describe('pure <PureEditableText>', () => {
             target: { value: 'Bar' },
         });
         expect(input.prop('value')).toBe('Bar');
+    });
+
+    it('notifies edit end on input blur', () => {
+        const handleEditEnd = jest.fn();
+        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
+        const input = wrapper.find('input');
+
+        input.simulate('blur', { target: { value: 'Foo' } });
+        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Foo', reset: false });
+    });
+
+    it('notifies edit end on Enter key', () => {
+        const handleEditEnd = jest.fn();
+        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
+        const input = wrapper.find('input');
+
+        input.simulate('keydown', {
+            target: { value: 'Bar' },
+            keyCode: keycode('Enter'),
+        });
+        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Bar', reset: false });
+        expect(handleEditEnd).toHaveBeenCalledTimes(1);
+    });
+
+    it('notifies edit end and reminds to reset on Escape key', () => {
+        const handleEditEnd = jest.fn();
+        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
+        const input = wrapper.find('input');
+
+        input.simulate('keydown', {
+            target: { value: 'Bar' },
+            keyCode: keycode('Escape'),
+        });
+        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Bar', reset: true });
+        expect(handleEditEnd).toHaveBeenCalledTimes(1);
+    });
+
+    it('prevents another notifiying edit end within short time interval', async () => {
+        const handleEditEnd = jest.fn();
+        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
+        const input = wrapper.find('input');
+
+        input.simulate('keydown', {
+            keyCode: keycode('Escape'),
+        });
+        expect(handleEditEnd).toHaveBeenCalledTimes(1);
+
+        await delayForMilliseconds(100);
+
+        input.simulate('blur');
+        expect(handleEditEnd).toHaveBeenCalledTimes(1);
+    });
+
+    it('does not interupt user input on non-Enter/Esc keys', () => {
+        const handleEditEnd = jest.fn();
+        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
+        const input = wrapper.find('input');
+
+        input.simulate('keydown', {
+            target: { value: 'Foo' },
+            keyCode: keycode('B'),
+        });
+        expect(handleEditEnd).not.toHaveBeenCalled();
+    });
+
+    it('can manually focus its <input> on DOM', () => {
+        const wrapper = mount(<PureEditableText />);
+        const input = wrapper.instance().inputNode;
+
+        const spy = jest.spyOn(input, 'focus');
+        wrapper.instance().focusInputNode();
+
+        expect(spy).toHaveBeenCalled();
     });
 });